feat: configurable permissions (#51)
* feat: configurable permissions

* docs: update docs for this new feature

* chore: update permissions to what gori requested

* refactor: make this pr simpler by using existing config file

there was a comment instructing to not edit the config file manually,
but I couldn't find a good reason for that, as the file is not generated
by any other tool, and used directly. Probably the comment meant to
express that configs from that file will apply to _all_ created repos,
and to not modify it for a particular workflow run.

* chore: run linters

* docs: explain how to disable a contentious feature